+++ tests/Transport/MimeTypeResolver/ApacheMimeTypeResolverTest.php
@@ -4,7 +4,8 @@ declare(strict_types=1);
 
 namespace Phptg\BotApi\Tests\Transport\MimeTypeResolver;
 
-use HttpSoft\Message\Stream;
+use Phptg\BotApi\Transport\InputFileData;
+use Phptg\BotApi\Transport\ResourceReader\NativeResourceReader;
 use PHPUnit\Framework\Attributes\DataProvider;
 use PHPUnit\Framework\TestCase;
 use Phptg\BotApi\Transport\MimeTypeResolver\ApacheMimeTypeResolver;
@@ -16,20 +17,19 @@ final class ApacheMimeTypeResolverTest extends TestCase
 {
     public static function dataBase(): iterable
     {
-        yield [null, new InputFile(new Stream())];
-        yield [null, new InputFile(new Stream(), 'test.non-exist-extension')];
-        yield ['image/jpeg', new InputFile(new Stream(), 'test.jpg')];
-        yield ['image/jpeg', new InputFile(new Stream(), 'TEST.JPG')];
+        yield [null, InputFile::fromLocalFile(__DIR__ . '/files/test.unknown')];
         yield ['text/plain', InputFile::fromLocalFile(__DIR__ . '/files/test.txt')];
         yield ['text/css', InputFile::fromLocalFile(__DIR__ . '/files/test.txt', 'test.css')];
+        yield ['text/css', InputFile::fromLocalFile(__DIR__ . '/files/test.txt', 'test.CSS')];
     }
 
     #[DataProvider('dataBase')]
     public function testBase(?string $expected, InputFile $file): void
     {
         $resolver = new ApacheMimeTypeResolver();
+        $fileData = new InputFileData($file, [new NativeResourceReader()]);
 
-        $result = $resolver->resolve($file);
+        $result = $resolver->resolve($fileData);
 
         assertSame($expected, $result);
     }