+++ tests/Transport/MimeTypeResolver/CustomMimeTypeResolverTest.php
@@ -4,7 +4,9 @@ declare(strict_types=1);
 
 namespace Phptg\BotApi\Tests\Transport\MimeTypeResolver;
 
-use HttpSoft\Message\Stream;
+use Phptg\BotApi\Tests\Support\StubResourceReader;
+use Phptg\BotApi\Transport\InputFileData;
+use Phptg\BotApi\Transport\ResourceReader\NativeResourceReader;
 use PHPUnit\Framework\Attributes\DataProvider;
 use PHPUnit\Framework\TestCase;
 use Phptg\BotApi\Transport\MimeTypeResolver\CustomMimeTypeResolver;
@@ -16,12 +18,11 @@ final class CustomMimeTypeResolverTest extends TestCase
 {
     public static function dataBase(): iterable
     {
-        yield [null, new InputFile(new Stream())];
-        yield [null, new InputFile(new Stream(), 'test.non-exist-extension')];
-        yield ['image/jpeg', new InputFile(new Stream(), 'test.jpg')];
-        yield ['image/jpeg', new InputFile(new Stream(), 'TEST.JPG')];
+        yield [null, new InputFile(null)];
+        yield [null, InputFile::fromLocalFile(__DIR__ . '/files/test.unknown')];
         yield ['text/plain', InputFile::fromLocalFile(__DIR__ . '/files/test.txt')];
         yield ['text/css', InputFile::fromLocalFile(__DIR__ . '/files/test.txt', 'test.css')];
+        yield ['text/css', InputFile::fromLocalFile(__DIR__ . '/files/test.txt', 'test.CSS')];
     }
 
     #[DataProvider('dataBase')]
@@ -32,8 +33,15 @@ final class CustomMimeTypeResolverTest extends TestCase
             'txt' => 'text/plain',
             'css' => 'text/css',
         ]);
+        $fileData = new InputFileData(
+            $file,
+            [
+                new NativeResourceReader(),
+                new StubResourceReader(uri: 'custom://test'),
+            ],
+        );
 
-        $result = $resolver->resolve($file);
+        $result = $resolver->resolve($fileData);
 
         assertSame($expected, $result);
     }